+++ Lib/fontbakery/profiles/googlefonts.py
@@ -25,7 +25,6 @@ PROFILE = {
             "com.google.fonts/check/metadata/gf_axisregistry_bounds",
             "com.google.fonts/check/metadata/gf_axisregistry_valid_tags",
             "com.google.fonts/check/metadata/has_regular",
-            "com.google.fonts/check/metadata/has_tags",
             "com.google.fonts/check/metadata/includes_production_subsets",
             "com.google.fonts/check/metadata/single_cjk_subset",
             "com.google.fonts/check/metadata/license",